fix(12): address Cubic code review violations in codec layer
P1 fixes:
- LosslessHuffman: correct BITS/VALUES mismatch (sum 17 not 19)
- JpegLosslessDecoder: add segment length bounds validation
- JpegLosslessDecoder: validate SOS marker presence before decode
- J2kEncoder: handle signed vs unsigned 32-bit samples correctly
- JpegBaselineDecoder: validate SOF0 before processing SOS
- HuffmanTable: fix end-of-data detection in bit reader

P2 fixes:
- J2kCodestream: validate code-block exponents (0-8 range)
- DwtTransform: use long multiplication to avoid overflow
- BitWriter: validate bit count lower bound (must be >= 1)
- JpegCodecOptions: presets return new instances to prevent mutation
- QuantizationTable: use long arithmetic in CreateScaled
- Jpeg2000CodecTests: use larger image for reliable compression test

+++ src/SharpDicom/Codecs/Jpeg/QuantizationTable.cs
@@ -255,9 +255,10 @@ namespace SharpDicom.Codecs.Jpeg
 
             for (int i = 0; i < 64; i++)
             {
-                // Scale and clamp to 1-255 (8-bit) or 1-65535 (16-bit)
-                int scaled = (baseValues[i] * scale + 50) / 100;
-                values[i] = Math.Max(1, Math.Min(255, scaled));
+                // Scale and clamp to 1-255 for 8-bit precision
+                // Use long arithmetic to avoid overflow with extreme scale values
+                long scaled = ((long)baseValues[i] * scale + 50) / 100;
+                values[i] = (int)Math.Max(1, Math.Min(255, scaled));
             }
 
             return new QuantizationTable(values, 8, tableId);
