+++ src/SharpDicom/Codecs/Jpeg2000/Wavelet/DwtTransform.cs
@@ -35,7 +35,8 @@ namespace SharpDicom.Codecs.Jpeg2000.Wavelet
                 throw new ArgumentException("Width and height must be positive.");
             }
 
-            if (data.Length < width * height)
+            // Use long multiplication to avoid overflow before comparing
+            if (data.Length < (long)width * height)
             {
                 throw new ArgumentException("Data buffer is too small for the specified dimensions.");
             }
@@ -86,7 +87,8 @@ namespace SharpDicom.Codecs.Jpeg2000.Wavelet
                 throw new ArgumentException("Width and height must be positive.");
             }
 
-            if (data.Length < width * height)
+            // Use long multiplication to avoid overflow before comparing
+            if (data.Length < (long)width * height)
             {
                 throw new ArgumentException("Data buffer is too small for the specified dimensions.");
             }
